requirements.txt
django==3.2.7
django-environ==0.7.0
psycopg2==2.9.1
gunicorn==20.1.0